Add Tooltips
Whenever we show the burned, consumed or difference energy, we now also
show a tooltip. This was done so that we can show the active and basal
energy for the burned energy.

This also fixes the tooltip for the bar chart, which contained an
unnecessary new line after we displayed the difference.