Display All Data in Bar Chart
It is now possible to display all the available data in the bar chart.
Until now we only displayed the difference between the burned and
consumed energy in the bar chart, now it is also possible to show the
burned or consumed energy. The value can be selected via a popup menu
similar to the selection of the time frame.