Fix Displayed Data for Current Day
Instead of the `getHealthDataFromTypes` function we are now using the
`getHealthIntervalDataFromTypes` function from the History widget to get
the data in the today widget. For that we have to adjust the interval to
only contain the data for the current day.

This was done because the `getHealthDataFromTypes` function returns
other values then the `getHealthIntervalDataFromTypes` function, which
are incorrect when they are compared with the values in Apple Health.

See https://github.com/cph-cachet/flutter-plugins/issues/982